# REZ-api-gateway/src/openapi.yaml
openapi: 3.0.3
info:
  title: REZ Intelligence API
  description: |
    AI-powered platform for local commerce with 3 client types:
    - REZ_ECOSYSTEM: Full intelligence sharing
    - NON_REZ: Strict tenant isolation
    - RABTUL_SAAS: Plugin-based SaaS
  version: 1.0.0
  contact:
    name: REZ Engineering
    email: engineering@rez.money
    url: https://rez.money
servers:
  - url: http://localhost:4300
    description: Local development
  - url: https://rez-api-gateway.rez.money
    description: Production
...
paths:
...
  /api/workflows:
    get:
      tags: [Workflows]
      summary: List workflows
      security:
        - ApiKeyAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Workflow list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowList'

    post:
      tags: [Workflows]
      summary: Create workflow
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkflowRequest'
      responses:
        '201':
          description: Workflow created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
  /api/workflows/{workflowId}:
    get:
      tags: [Workflows]
      summary: Get workflow
      security:
        - ApiKeyAuth: []
      parameters:
        - name: workflowId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Workflow details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key for tenant authentication.
        Format: {prefix}_{tenant}_{id}
        - REZ_ECOSYSTEM: rez_*
        - NON_REZ: ext_*
        - RABTUL_SAAS: saas_*

    InternalToken:
      type: apiKey
      in: header
      name: X-Internal-Token
      description: Internal service-to-service authentication
  schemas:
...
    Workflow:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [draft, active, paused, archived]
        entryNodeId:
          type: string
        nodeCount:
          type: integer
        executionCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    WorkflowList:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            workflows:
              type: array
              items:
                $ref: '#/components/schemas/Workflow'
            total:
              type: integer
            page:
              type: integer
            limit:
              type: integer
    CreateWorkflowRequest:
      type: object
      required: [name, entryNodeId, nodes]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        entryNodeId:
          type: string
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowNode'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowEdge'
    WorkflowNode:
      type: object
      required: [id, type, config]
      properties:
        id:
          type: string
        type:
          type: string
          enum:
            - action_send_email
            - action_send_sms
            - action_send_whatsapp
            - action_send_push
            - action_add_coins
            - action_deduct_coins
            - action_create_order
            - condition_if_time
            - condition_if_segment
            - delay_wait_hours
            - ai_analyze_sentiment
            - ai_generate_response
        config:
          type: object
          properties:
            actionType:
              type: string
            params:
              type: object
    WorkflowEdge:
      type: object
      properties:
        from:
          type: string
        to:
          type: string
        condition:
          type: string
...